Refactor: Modification du bean localisation, ajout de l'url d'une image et d'un uid
